+++ Project.toml
@@ -11,6 +11,18 @@ Requires = "ae029012-a4dd-5104-9daa-d747884805df"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 Unicode = "4ec0a83e-493e-50e2-b9ac-8f72acf5a8f5"
 
+[weakdeps]
+JSON = "682c06a0-de6a-54ab-a142-c8b1cf79cde6"
+RecipesBase = "3cdcf5f2-1ef4-517c-9805-6587b60abb01"
+SentinelArrays = "91c51154-3ec4-41a3-a24f-3f23e20d615c"
+StructTypes = "856f2bd8-1eba-4b0a-8007-ebc267875bd4"
+
+[extensions]
+CategoricalArraysJSONExt = "JSON"
+CategoricalArraysRecipesBaseExt = "RecipesBase"
+CategoricalArraysSentinelArraysExt = "SentinelArrays"
+CategoricalArraysStructTypesExt = "StructTypes"
+
 [compat]
 DataAPI = "1.6"
 JSON = "0.15, 0.16, 0.17, 0.18, 0.19, 0.20, 0.21"
